small changes for manuscript
Corrected a mistake in total proportion of variation for Tab. S1 (for 7-factor solution, the values was derived from the last factor listed which was not necessarily the last one derived because I reordered them beforehand), changed formatting of tables such that they are easier to import to word (got rid of quotes) and reformatted some plots such that the text is easier to read (based on how they looked in the manuscript).